requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.25.0
pyyaml>=6.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
questionary>=2.0.0
